remove duck_array_ops  module
